requirements.txt
numpy>=1.23.0,<2.0.0
scipy>=1.10.0,<2.0.0
PyYAML>=6.0
matplotlib>=3.7.0,<4.0.0
pytest>=7.0.0
streamlit>=1.34.0,<2.0.0
requests>=2.31.0,<3.0.0
pandas>=1.5.0,<3.0.0
altair>=5.0.0,<6.0.0
